requirements.txt
langchain>=0.1.0
transformers
sentence-transformers
requests
python-dotenv
beautifulsoup4
torch 